requirements.txt
sympy~=1.5.1
pandas~=0.24.2
numpy~=1.16.4
requests~=2.22.0
aiohttp~=2.3.10
APScheduler~=3.6.3
websocket~=0.2.1
websocket-client~=0.56.0
pysocks